+++ Runtime/Scripts/PrimitiveCreateContextBase.cs
@@ -7,6 +7,8 @@ namespace GLTFast {
     abstract class PrimitiveCreateContextBase {
         public int primtiveIndex;
         public int[] materials;
+        public bool needsNormals;
+        public bool needsTangents;
         public abstract bool IsCompleted {get;}
         public abstract Primitive? CreatePrimitive();
     }