+++ Runtime/Scripts/PrimitiveDracoCreateContext.cs
@@ -38,13 +38,13 @@ namespace GLTFast {
             var mesh = DracoMeshLoader.CreateMesh(dracoMesh, out hasNormals, out hasTexcoords);
             Profiler.EndSample();
 
-            if(!hasNormals) {
+            if(needsNormals && !hasNormals) {
                 Profiler.BeginSample("Draco.RecalculateNormals");
                 // TODO: Make optional. Only calculate if actually needed
                 mesh.RecalculateNormals();
                 Profiler.EndSample();
             }
-            if(hasTexcoords) {
+            if(needsTangents && hasTexcoords) {
                 Profiler.BeginSample("Draco.RecalculateTangents");
                 // TODO: Make optional. Only calculate if actually needed
                 mesh.RecalculateTangents();