+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Normals and tangents (if not present) are only calculated if the assigned material actually requires them.
+
 ## [0.10.1] - 2020-02-24
 ### Added
 - Experimental KTX / Basis Universal support was merged (off by default)